+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog for NeoFS Node
 ### Changed
 
 - Allow to evacuate shard data with `EvacuateShard` control RPC (#1800)
+- Flush write-cache when moving shard to DEGRADED mode (#1825)
 
 ### Fixed
 - Description of command `netmap nodeinfo` (#1821)